+++ connector/syncthingdev.h
@@ -33,8 +33,8 @@ struct LIB_SYNCTHING_CONNECTOR_EXPORT SyncthingDev
     int progressRate = 0;
     bool introducer = false;
     bool paused = false;
-    int totalIncomingTraffic = 0;
-    int totalOutgoingTraffic = 0;
+    uint64 totalIncomingTraffic = 0;
+    uint64 totalOutgoingTraffic = 0;
     QString connectionAddress;
     QString connectionType;
     QString clientVersion;
